refactoring file structure temporarily
